requirements.txt
Authlib==0.15.2
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
dnspython==1.16.0
Flask==1.1.2
Flask-Limiter==1.4
Flask-Login==0.5.0
Flask-PyMongo==2.3.0
gunicorn==20.0.4
httpie==2.3.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
limits==1.5.1
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==8.0.1
pycparser==2.20
Pygments==2.7.3
PyJWT==1.7.1
pymongo==3.11.2
pyOpenSSL==20.0.0
PySocks==1.7.1
python-bsonjs==0.2.0
python-dotenv==0.15.0
requests==2.25.0
requests-toolbelt==0.9.1
six==1.15.0
urllib3==1.26.2
uuid==1.30
Werkzeug==1.0.1
